refactor: Separate WorkRequest Constant
